requirements.txt
aniso8601==7.0.0
asgiref==3.6.0
attrs==23.1.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.33.0
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-better-admin-arrayfield==1.4.2
django-cloudflare-images==0.5.2
django-cors-headers==4.0.0
django-filter==23.1
django-import-export==3.2.0
django-map-widgets==0.4.1
django-modeltranslation==0.18.10
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
drf-jwt==1.19.2
drf-spectacular==0.26.3
et-xmlfile==1.1.0
flake8==6.0.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphene-gis==0.0.7
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.1.3
martor==1.6.26
mccabe==0.7.0
odfpy==1.4.1
openpyxl==3.1.2
phonenumbers==8.13.14
Pillow==10.0.0
promise==2.3
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
random-username==1.0.2
requests==2.28.2
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
text-unidecode==1.3
typing_extensions==4.7.0
uritemplate==4.1.1
urllib3==1.26.15
validators==0.22.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
